Conflicts:
	src/game/Level3.cpp
	src/game/SpellEffects.cpp
	src/game/Transports.cpp
	src/game/Unit.cpp

+++ src/game/AntiCheat.cpp
@@ -366,7 +366,7 @@ bool AntiCheat::CheckNeeded(AntiCheatCheck checktype)
         || !GetPlayer()->IsInWorld()
         || GetPlayer()->IsBeingTeleported()
         || GetPlayer()->GetPlayerbotAI()
-        || GetPlayer()->GetSession()->GetSecurity() > sWorld.getConfig(CONFIG_UINT32_ANTICHEAT_GMLEVEL))
+        || GetPlayer()->GetSession()->GetSecurity() > int32(sWorld.getConfig(CONFIG_UINT32_ANTICHEAT_GMLEVEL)))
         return false;
 
     if (GetMover()->HasAuraType(SPELL_AURA_MOD_CONFUSE))
